Windows: Adds support for Host Compute Calls for Win-Bridge
HNS is moving to a new API, this the first of a couple pull requests
that will move the windows plugin to use the Host Compute (v2) APIs, as well
as clean-up the code and improve testing coverage. This commit in particular
moves the win-bridge to use either the old API or Host Compute (v2) api
depending on a conf parameter.